Fix position not being updated correctly bug
Fixed a bug where the position of the player in the Game component was
actually their previous position when they moved, thus their position in
the world was always a step behind their actual position. This was due
to the redux sagas being async. Fixed the issue by catching the change
in state within the componentWillReceiveProps React Component method,
and only when the new position has been received do we re-render the
canvas to display their new position.